Make all primitives return a value
All primitives now leave a value on the stack. The ones that used to not
return anything (printn for example) now return nil. This way, we don't
have to check if the stack is empty or not when executing "ret"
instruction.

The ng machine now supports a new instruction: "drop". This simply drops
the value on top of the stack. The Lisp compiler puts drop instructions
after all forms except the last one in a lambda body.