Add "define" and multiple top-level forms support
We can now have multiple top-level forms. The compiler emits a drop
instruction after all of them except for the last.

We also added the "define" form which adds a variable to the current
environment and sets it to the given value. In order to be able to do
that, we've also added two new instructions: "xp" which expands the
current (first) frame in the environment by one, and "st" which acts
like the counterpart to "ld" and stores a value from the stack to the
given location.