Add rest parameter
The rest parameter is the last parameter of a lambda which must be
preceded with the & symbol. For example in (lambda (x y & z) z) the z
parameter is a rest parameter. It captures any extra arguments passed to
the function after the first two.

In order to support this feature, a new instruction was added to the
SECD machine named ldfx which is the expanded form of ldf. This
instructions receives a list of size two. The first value in the list is
an integer (in range 0 255) which tells the number of non-rest arguments
the function receives. The second value is the same as the ldf operand.

"ap" and "tap" instructions support closures created using ldfx. These
instructions would pass any extra arguments as the last argument.

Notice that the "rap" instruction does not support closures created with
ldfx.