Add defmac
defmac allows defining macros. Macros are only allowed to be defined at
the top-level.

The way we implement macros is by holding the code generated for all
top-level "define" forms (after macro expansion). Any defmac is
converted to a macro object which also captures the environment.

When we want to expand macros, we create an Secd machine and load the
top-level code into it. We also convert the macro body to a function and
call it with the passed (unevaluated) arguments. We then run the
machine, and gather the results which is used as macro expansion.