Revamp define and add set!
We want functions defined using the "define" form to be able to call any
other functions even if defined later. So now, if "define" is used at
the top-level, instead of expanding the top-level environment, we
associate a value with the symbol. This associated value can be read at
run-time too.

So the top-level environment is also removed. The two new instructions
"set" and "get" get a symnum and write/read the symbol's associated
value.

We also added the set! primitive form. set! sets a symbol value or sets
a local variable, depending on the symbol.

Notice that we now have the "st" instruction for setting local
variables, and "set" for setting symbols. That's certainly not ideal
naming!