Add error primitive form/instruction
The error form receives at list a symbol as error type. It might also
get a list of key-values afterwards as metadata for the error, for
example:

(error :file-not-found
       :msg "File not found"
       :filename "foo.txt")

This is interpreted as an error of type :file-not-found with two
metadata values named :msg and :filename.

The :msg metadata could be interpreted specially and printed to the
terminal if the error happens.