Fix gensyms
The way gensyms worked before did not work correctly. We were assuming
that gensyms never get to the assembly code, but if a macro uses another
macro in its expansion, and it passes the inner macro a gensym, then the
gensym will be passed to the assembler as part of the inner macro's
input (which is unevaluated).

New gensyms can be written and read. They are basically just uuids, that
also have a short form so they look better in printed form.

Notice that right now if we print a gensym and read it again, the short
form is lost. This can be fixed relatively easily by parsing the unique
name (which already includes a short counter) and expand that into a
short name, but we don't do that for now (because I'm lazy).