Bring duplicate "define" behavior closer to R7RS
Global defines now behave as R7RS requires: if a we define a name that
is already defined, we simply set it to the new value.

Local defines are still not entirely standards-compliant. First they
should be only allowed at the beginning of a body, and second, while
shadowing local variables is allowed, duplicate definitions should be an
error.

So this is allowed:

(let ((x 100))
  (define x 200)
  (print x))

But this is an error:

(let ((x 100))
  (define y 200)
  (define y 300)
  (print y))