Add non-type-checking versions of stack operations
Sometimes we're already sure of what we are pushing on the stack, for
example:

 - we're pushing a value we just created: self.s.push(Nil())
 - we're pushing a value we already received using a typed pop
   operation.
 - we've already performed type-checking; for example, Pair already
   performs type-checking on its car and cdr, so no reason to type check
   again when pushing car/cdr of a pair.

Similarly for top/pop we might not care about the type. For all these
cases, we've added versions of push/pop/top operations, with an x added
to the end (pushx/topx/popx) that do not perform type checking on the
values.