Make macros receive the same code objects passed
Until now, macros would generate code to construct passed arguments on
the stack and pass them to the macro function.

The trouble with this approach was that the quoted code blocks passed to
the macro were not the same objects as the ones originally passed, so
any source location data attached to these blocks of code would be lost
in the output.

We now use a different approach: the arguments are directly pushed to
the machine's stack, and a "get" instruction followed by an "ap"
instruction perform the function call. This means the functions receive
exactly the same objects as constructed by the compiler, and if they are
used in the output, they still have their source location data attached.