Fix implementation of case-lambda
This was a double bug, with one bug canceling the other. The code
generated by case-lambda was checking to see if the argument list is
improper, which it of course never is, but the is-improper function was
actually checking if the list is proper!

What's more, we wouldn't need an is-improper function in any case. list?
already checks for proper lists.