Do argument checking before adding continuation
Before this, when there was an argument mismtach, there was an extra
stack frame shown in the error report because we've already added a
continuation for the call (that had not happened yet). The shown stack
frame was bugged and confusing of course.

We now perform the argument checking/fitting before adding the
continuation in order to prevent that.

This also solves another potential issue. The values we add to the "e"
register should be python lists, but before we call "fit_args" they're a
scheme list. In the previous state, "rap" would add "args" to "e" before
it was converted to a python list by "fit_args".