Let unhandled exceptions go back to SECD as errors
So far unhandled exceptions were handled by stdlib, which would print an
error and call exit. This meant that for example we'd never get a stack
trace, neither for system errors nor for user errors.

We now have the built-in "abort" which allows exceptions to be sent back
to the SECD machine as errors. We also keep track of the original
continuation in case of system errors, so we'd get the exact same stack
trace as before.