Add local macros
We can now use define-macro in a let/lambda body, and the macro will
then be available only inside that body.

Notice that the macro does NOT have access to the local variables in its
own code. For example the following causes an error (invalid frame
number):

    (let ((x 10))
      (define-macro (foo a)
        `(list ,a ,x))
      (print (foo 99)))

This would try to read the x variable as part of the macro body, but the
macro is not really running in an environment where x is bound.

Notice that this would work however:

    (let ((x 10))
      (define-macro (foo a)
        `(list ,a x))
      (print (foo 99)))

Here x is returned as a symbol which is read after expansion.