Improve "undefined symbol" errors
Suppose we're missing the import for the core library, and there's a
definition like this:

    (define x 100)

Before this change, we'd get an error saying "symbol x is read but never
defined". The actual error is that we're compiling "define" as a
function rather than a special form.

Now we compile the function code first, so if there's an undefined
symbol here, it is raised before any "undefined symbol" errors in
arguments.

Notice that this still won't work with other kinds of errors, because we
raise "undefined" errors at the end of compilation, while others are
raised immediately.