Fix error during compile error
If we wanted to raise an error about compiling an improper list, we'd
run this line:

    form = form or self.form

Trouble is, python would try to call __len__ on form to see if its
length is zero or not (and see if it evaluates to true or not). We now
explicitly compare to None.