Fix "constructed" expressions
The compiler would construct new expressions out of existing ones and
have those compiled. For example, it converts "let" into a "lambda" call
or a "letrec". We were using literal symbols to do so which might not
always work since for example the special form "lambda" might not be
available in a certain environment, or it might be renamed to something
else.

The solution here is to allow special forms to be used with the "#$"
prefix just like primcalls. These will always be available with their
original meaning.