Refactor and fix Environment
We now have a ToplevelEnvironment and a LocalEnvironment. Local
environments each hold a single frame, and are chained together until
they reach a top-level environments.

This will allow fixing hygienic macros to work with local variables
later, but it also fixes a few existing issues. First of all, the copy
method had to be maintained and if we added fields to Environment
without handling it in copy, it could lead to subtle and hard to track
bugs.

Also, the top-level information like defined_symbols and
read_free_symbols should be shared between different environments
derived from one, but we were actually losing future changes to the
top-level after copy, and vice versa.

One downside to the new design however is that if any method is added to
ToplevelEnvironment, it likely needs to be added to LocalEnvironment
too, with a bit of copy-paste code to call top-level's version. But at
least in this case if we forget to do it, we probably run into immediate
errors.