Allow patterns to match any cdr of a list
A pattern (x y . z) can now match any proper or improper lists with at
least three items. It would match, for example, (1 2 3 4 5), and would
set x to 1, y to (2 3 4), and z to (5).

Notice that this won't work with lists that have an ellipsis inside
them, since the ellipsis would try to match any extra items. This is
correct behavior, if I'm understanding R7RS correctly.