Allow macros to refer to symbols not yet defined
It is allowed to refer to a symbol that is going to be defined later in
the same environment, as long as the reference is not immediately
executed (that is, inside a procedure).

However, up until now, if a macro referred to a name that was to be
defined after it's definition, we would record the current meaning of
that symbol (that is, FREE) and keep using that, even if the symbol
would become defined later.

As an example, while this was allowed:

   (define (foo) x)
   (define x 100)

This was not:

   (define-syntax m
     (syntax-rules ()
       ((_) x)))
   (define (foo) (m))
   (define x 100)

or not even this:

   (define-syntax m
     (syntax-rules ()
       ((_) x)))
   (define x 100)
   (define (foo) (m))

This is now fixed.