Fixed nan comparison
nan is neither equal to, nor greater than, nor smaller than any other
number. so our previous approach of defining "greaten than" to be the
inverse of "less than" did not work for it.

We could check for nan, but I decided to just add two other instructions
to the vm for comparisons.