Refactor exceptions
All exceptions that have source location info (.form and .source
attributes) now sub-class CompileError. This means we don't need to
always explicitly catch and re-raise those as CompileError. We still do
that in some places were we want to attach more information, like a
better error message, or a more specific form to point to.