Change the semantics of apply instructions
Instead of accepting a list, ap, tap, and rap instructions now have an
operand which says how many arguments are passed, and then at runtime
the arguments are read one by one from the stack.

For example, (foo 100 200 300) is compiled as:

    (ldc 300 ldc 200 ldc 100 get foo ap 3)

This reduces the number of runtime conses considerably. We also have a
new "apx" instruction which accepts a list of arguments, and is used to
implement the "apply" procedure.