Make sel/join instructions not use the dump
This makes the sel/join instructions behave similar to how normal cpus
implement the conditionals. the sel instruction receives an offset to
skip when the condition is false. the join instruction also receives an
output and behaves exactly like a jump instruction.

The true body of the sel instruction is expected to end with a join as
normal, but false is expected not to end in a join.

The assembler adds the size of true body after the sel instruction, and
the size of the false body at the end of the true body (and thus after
the join instruction).

This has two important consequences:

 - Tail call elimination now works correctly. Previously, when the tail
   call was inside an "if", it would still add a continuation to the
   dump (the one added by the "if" itself).

 - Performance of conditionals has been improved greatly. The test suite
   now runs at about half the time as before.