Fix bound identifiers matching macro literals
In:

(let ((else #f))
  (cond (else (display "oops!"))))

The symbols 'else' should not match the literal else, and so the
expression display "oops!".

At least, this is according to Chez Scheme docs and implementation. See:
https://www.scheme.com/tspl2d/syntax.html

Note that some other implementations (like Gambit and Chicken) don't
work like this, but I think Chez Scheme has the correct implementation
in this case.

I can't find anything in R7RS to disambiguate this, but I think R6RS
does mention somewhere that literals are matched using
free-identifier=?.