Fix io/cpeek runtime call
What we were doing was very incorrect since we were reading the
underlying buffer in peek and the file itself in read, which could cause
issues since the read position in the two could be different.

We now simply use tell/seek which is much more robust and should be
available on the files we expect here.