Fix dummy frame handling
There was a serious issue with how dummy frames were used. It would
cause the following code to fail, because of accessing the dummy frame:

    (define (foo) 1)

    (letrec ((port (foo))
             (return-value (lambda () port)))
      (return-value))

The reason for failure is that the even though the `rap` instruction
overwrites the dummy frame, the lambda still refers to it in its
environment.

We now handle this very differently:

The `dum` instruction creates a new instance of a `Frame` class and adds
it to the environment. This object is at first a "dummy" frame and
accessing it (like a list) would cause an error. But when `rap` is
called, it sets the new frame data inside this object and from then on,
it acts like a normal list with the new data.

This way, whoever has gotten a reference of a dummy frame, will see the
changes when the it is replaced by `rap`.