Add map circular list tests
Turns out that our map actually does work correctly with cicular tests.
The spec says: The lists can be circular, but it is an error if all of
them are circular. So one non-circular list is enough to terminate the
execution, and that already works with our implementation.

I contemplated adding a test to check if all input lists are circular
and raise an error. We could use the circular-list? function from srfi
1. But that would add an overhead to all map calls, and the spec does
not require us to raise an error.